+++ src/Fl_File_Input.cxx
@@ -120,7 +120,7 @@ void Fl_File_Input::update_buttons() {
 
     end ++;
 
-    buttons_[i] = (short)fl_width(start, end - start);
+    buttons_[i] = (short)fl_width(start, (int) (end - start));
     if (!i) buttons_[i] += Fl::box_dx(box()) + 6;
   }
 
@@ -273,7 +273,7 @@ Fl_File_Input::handle_button(int event)		// I - Event
   if (i < 0) {
     // Found the end; truncate the value and update the buttons...
     *start = '\0';
-    value(newvalue, start - newvalue);
+    value(newvalue, (int) (start - newvalue) );
 
     // Then do the callbacks, if necessary...
     set_changed();
