+++ src/Fl_PostScript.cxx
@@ -935,7 +935,7 @@ void Fl_PostScript_Graphics_Driver::font(int f, int s) {
   Fl_Font_Descriptor *desc = driver->font_descriptor();
   this->font_descriptor(desc);
   if (f < FL_FREE_FONT) {
-    float ps_size = s;
+    float ps_size = (float) s;
     fprintf(output, "/%s SF\n" , _fontNames[f]);
 #if defined(USE_X11) 
 #if USE_XFT
