+++ src/Fl_Preferences.cxx
@@ -741,7 +741,7 @@ char Fl_Preferences::set( const char *key, const char *text ) {
 
 // convert a hex string to binary data
 static void *decodeHex( const char *src, int &size ) {
-  size = strlen( src )/2;
+  size = (int) strlen( src )/2;
   unsigned char *data = (unsigned char*)malloc( size ), *d = data;
   const char *s = src;
   for ( int i=size; i>0; i-- ) {
@@ -848,7 +848,7 @@ char Fl_Preferences::set( const char *key, const void *data, int dsize ) {
  */
 int Fl_Preferences::size( const char *key ) {
   const char *v = node->get( key );
-  return v ? strlen( v ) : 0 ;
+  return (int) (v ? strlen( v ) : 0);
 }
 
 /**
@@ -956,7 +956,7 @@ static char makePath( const char *path ) {
   if (access(path, 0)) {
     const char *s = strrchr( path, '/' );
     if ( !s ) return 0;
-    int len = s-path;
+    size_t len = s-path;
     char *p = (char*)malloc( len+1 );
     memcpy( p, path, len );
     p[len] = 0;
@@ -997,7 +997,7 @@ Fl_Preferences::RootNode::RootNode( Fl_Preferences *prefs, Root root, const char
 #ifdef WIN32
 #  define FLPREFS_RESOURCE	"Software\\Microsoft\\Windows\\CurrentVersion\\Explorer\\Shell Folders"
 #  define FLPREFS_RESOURCEW	L"Software\\Microsoft\\Windows\\CurrentVersion\\Explorer\\Shell Folders"
-  int appDataLen = strlen(vendor) + strlen(application) + 8;
+  size_t appDataLen = strlen(vendor) + strlen(application) + 8;
   DWORD type, nn;
   LONG err;
   HKEY key;
@@ -1006,7 +1006,7 @@ Fl_Preferences::RootNode::RootNode( Fl_Preferences *prefs, Root root, const char
     case SYSTEM:
       err = RegOpenKeyW( HKEY_LOCAL_MACHINE, FLPREFS_RESOURCEW, &key );
       if (err == ERROR_SUCCESS) {
-        nn = FL_PATH_MAX - appDataLen; 
+        nn = (DWORD) (FL_PATH_MAX - appDataLen); 
         err = RegQueryValueExW( key, L"Common AppData", 0L, &type,
                                 (BYTE*)filename, &nn ); 
         if ( ( err != ERROR_SUCCESS ) && ( type == REG_SZ ) ) {
@@ -1019,7 +1019,7 @@ Fl_Preferences::RootNode::RootNode( Fl_Preferences *prefs, Root root, const char
     case USER:
       err = RegOpenKeyW( HKEY_CURRENT_USER, FLPREFS_RESOURCEW, &key );
       if (err == ERROR_SUCCESS) {
-        nn = FL_PATH_MAX - appDataLen;
+        nn = (DWORD) (FL_PATH_MAX - appDataLen);
         err = RegQueryValueExW( key, L"AppData", 0L, &type,
                                 (BYTE*)filename, &nn ); 
         if ( ( err != ERROR_SUCCESS ) && ( type == REG_SZ ) ) {
@@ -1043,7 +1043,7 @@ Fl_Preferences::RootNode::RootNode( Fl_Preferences *prefs, Root root, const char
     wcscpy(b, (xchar *) filename);
 #endif
     //  filename[fl_unicode2utf(b, wcslen((xchar*)b), filename)] = 0;
-    unsigned len = fl_utf8fromwc(filename, (FL_PATH_MAX-1), b, wcslen(b));
+    unsigned len = fl_utf8fromwc(filename, (FL_PATH_MAX-1), b, (unsigned) wcslen(b));
     filename[len] = 0;
     free(b);
   }
@@ -1158,17 +1158,17 @@ int Fl_Preferences::RootNode::read() {
   for (;;) {
     if ( !fgets( buf, 1024, f ) ) break;	// EOF or Error
     if ( buf[0]=='[' ) {			// read a new group
-      int end = strcspn( buf+1, "]\n\r" );
+      size_t end = strcspn( buf+1, "]\n\r" );
       buf[ end+1 ] = 0;
       nd = prefs_->node->find( buf+1 );
     } else if ( buf[0]=='+' ) {			// value of previous name/value pair spans multiple lines
-      int end = strcspn( buf+1, "\n\r" );
+      size_t end = strcspn( buf+1, "\n\r" );
       if ( end != 0 ) {				// if entry is not empty
 	buf[ end+1 ] = 0;
 	nd->add( buf+1 );
       }
     } else {					 // read a name/value pair
-      int end = strcspn( buf, "\n\r" );
+      size_t end = strcspn( buf, "\n\r" );
       if ( end != 0 ) {				// if entry is not empty
 	buf[ end ] = 0;
 	nd->set( buf );
@@ -1307,7 +1307,7 @@ int Fl_Preferences::Node::write( FILE *f ) {
     char *src = entry_[i].value;
     if ( src ) {		// hack it into smaller pieces if needed
       fprintf( f, "%s:", entry_[i].name );
-      int cnt, written;
+      size_t cnt, written;
       for ( cnt = 0; cnt < 60; cnt++ )
 	if ( src[cnt]==0 ) break;
       written = fwrite( src, cnt, 1, f );
@@ -1399,7 +1399,7 @@ void Fl_Preferences::Node::set( const char *line ) {
   } else {
     const char *c = strchr( line, ':' );
     if ( c ) {
-      unsigned int len = c-line+1;
+      size_t len = c-line+1;
       if ( len >= sizeof( nameBuffer ) )
         len = sizeof( nameBuffer );
       strlcpy( nameBuffer, line, len );
@@ -1415,8 +1415,8 @@ void Fl_Preferences::Node::set( const char *line ) {
 void Fl_Preferences::Node::add( const char *line ) {
   if ( lastEntrySet<0 || lastEntrySet>=nEntry_ ) return;
   char *&dst = entry_[ lastEntrySet ].value;
-  int a = strlen( dst );
-  int b = strlen( line );
+  size_t a = strlen( dst );
+  size_t b = strlen( line );
   dst = (char*)realloc( dst, a+b+1 );
   memcpy( dst+a, line, b+1 );
   dirty_ = 1;
@@ -1452,7 +1452,7 @@ char Fl_Preferences::Node::deleteEntry( const char *name ) {
 // - this method will always return a valid node (except for memory allocation problems)
 // - if the node was not found, 'find' will create the required branch
 Fl_Preferences::Node *Fl_Preferences::Node::find( const char *path ) {
-  int len = strlen( path_ );
+  int len = (int) strlen( path_ );
   if ( strncmp( path, path_, len ) == 0 ) {
     if ( path[ len ] == 0 )
       return this;
@@ -1494,9 +1494,9 @@ Fl_Preferences::Node *Fl_Preferences::Node::search( const char *path, int offset
 	return nn->search( path+2, 2 ); // do a relative search on the root node
       }
     }
-    offset = strlen( path_ ) + 1;
+    offset = (int) strlen( path_ ) + 1;
   }
-  int len = strlen( path_ );
+  int len = (int) strlen( path_ );
   if ( len < offset-1 ) return 0;
   len -= offset;
   if ( ( len <= 0 ) || ( strncmp( path, path_+offset, len ) == 0 ) ) {
